search modal component added
