--- Authentication/SignIn.js
@@ -1,9 +1,0 @@
-const SignIn = ({className}) => {
-    return (
-        <button className={className}>
-            Sign in
-        </button>
-    )
-}
-
-export default SignIn
