--- Authentication/SignUp.js
@@ -1,9 +1,0 @@
-const SignUp = ({className}) => {
-    return (
-        <button className={className}>
-            Sign Up
-        </button>
-    )
-}
-
-export default SignUp
