+++ app/layout.js
@@ -1,26 +1,20 @@
-import Navbar from '@/sections/Navbar'
-import './globals.css'
-import { Figtree } from 'next/font/google'
+import Navbar from "@/components/sections/Navbar";
+import "./globals.css";
+import { Figtree } from "next/font/google";
 
-const font = Figtree({ subsets: ['latin'] })
+const font = Figtree({ subsets: ["latin"] });
 
 export const metadata = {
-  title: 'BeatSocial',
-  description: 'listen to your Tune',
-}
+  title: "BeatSocial",
+  description: "listen to your Tune",
+};
 
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={font.className}>
-        
-        <Navbar>
-          {children}
-          </Navbar>
-         
-        
-
+        <Navbar>{children}</Navbar>
       </body>
     </html>
-  )
+  );
 }
