+++ app/page.js
@@ -1,23 +1,21 @@
-import SearchModal from "@/components/SearchModal";
-import Header from "@/sections/Header";
-import HomeContent from "@/sections/HomeContent";
+import Header from "@/components/sections/Header";
+import HomeContent from "@/components/sections/HomeContent";
 
 export default function Home() {
   return (
     <div
       className="
     
-    rounded-lg
     h-full
     w-full
     overflow-hidden
     overflow-y-auto
+    rounded-lg
     
     
         px-4
     "
     >
-      {/* <SearchModal /> */}
       <Header />
       <HomeContent />
     </div>
