+++ app/top-album/page.js
@@ -1,7 +1,21 @@
 export default function Page() {
-    return (
-        <>
-            <Content/>
-        </>
-    )
+  return (
+    <div
+      className="
+    
+    mt-2
+    h-full
+    w-full
+    overflow-hidden
+    overflow-y-auto
+        rounded-lg
+        px-4
+    "
+    >
+      <p className="mb-3 font-mono text-xl font-semibold text-secondary-orange">
+        Top Albums
+      </p>
+      <Content />
+    </div>
+  );
 }
