+++ components/SearchModal.js
@@ -2,28 +2,34 @@
 
 import { IoMdCloseCircleOutline } from "react-icons/io";
 
-const SearchModal = ({ openModal, isOpen }) => {
+const SearchModal = ({ openModal, isOpen, children }) => {
+  const handleClose = (e) => {
+    if (e.target.id === "wrapper") openModal();
+  };
   return (
     <>
       {isOpen ? (
         <div
+          id="wrapper"
           className="
   fixed
   inset-0
+  z-10
+  flex
+  cursor-pointer
+  items-center
+  justify-center
   bg-black
   bg-opacity-50
   backdrop-blur-sm
-  flex
-  justify-center
-  items-center
-  z-10
   "
+          onClick={handleClose}
         >
-          <div className="w-[600px] flex flex-col">
-            <button className="text-white" onClick={openModal}>
+          <div className="flex w-[600px] flex-col">
+            {/* <button className=" text-white" onClick={openModal}>
               <IoMdCloseCircleOutline size={24} />
-            </button>
-            <div className="bg-white p-4 rounded">our Modal</div>
+            </button> */}
+            <div className="rounded  p-2">{children}</div>
           </div>
         </div>
       ) : null}
