+++ components/auth/SignIn.js
@@ -1,0 +1,9 @@
+const SignIn = ({className}) => {
+    return (
+        <button className={className}>
+            Sign in
+        </button>
+    )
+}
+
+export default SignIn
