+++ components/auth/SignUp.js
@@ -1,0 +1,9 @@
+const SignUp = ({className}) => {
+    return (
+        <button className={className}>
+            Sign Up
+        </button>
+    )
+}
+
+export default SignUp
