+++ components/sections/LyricsSection.js
@@ -1,0 +1,9 @@
+const LyricSection = () => {
+    return (
+        <div>
+            the lyrics section
+        </div>
+    )
+}
+
+export default LyricSection
