+++ components/sections/PlayTrackSection.js
@@ -1,0 +1,9 @@
+const PlayTrackSection = () => {
+    return (
+        <div>
+            play/pause
+        </div>
+    )
+}
+
+export default PlayTrackSection
