+++ components/sections/SocialsSection.js
@@ -1,0 +1,9 @@
+const SocialsSection = () => {
+    return (
+        <div>
+            the socials section
+        </div>
+    )
+}
+
+export default SocialsSection
