+++ components/sections/TrackLineUpSection.js
@@ -1,0 +1,15 @@
+import PlayTrackSection from "./PlayTrackSection"
+import Track from "../Track"
+
+const TrackLineUpSection = () => {
+    return (
+        <div>
+            <Track/>
+            <Track/>
+            <Track />
+            <PlayTrackSection/>
+        </div>
+    )
+}
+
+export default TrackLineUpSection
