--- sections/LyricsSection.js
@@ -1,9 +1,0 @@
-const LyricSection = () => {
-    return (
-        <div>
-            the lyrics section
-        </div>
-    )
-}
-
-export default LyricSection
