--- sections/PlayTrackSection.js
@@ -1,9 +1,0 @@
-const PlayTrackSection = () => {
-    return (
-        <div>
-            play/pause
-        </div>
-    )
-}
-
-export default PlayTrackSection
