--- sections/SocialsSection.js
@@ -1,9 +1,0 @@
-const SocialsSection = () => {
-    return (
-        <div>
-            the socials section
-        </div>
-    )
-}
-
-export default SocialsSection
