--- sections/TrackLineUpSection.js
@@ -1,15 +1,0 @@
-import PlayTrackSection from "./PlayTrackSection"
-import Track from "../components/Track"
-
-const TrackLineUpSection = () => {
-    return (
-        <div>
-            <Track/>
-            <Track/>
-            <Track />
-            <PlayTrackSection/>
-        </div>
-    )
-}
-
-export default TrackLineUpSection
